pyproject.toml
[tool.poetry]
name = "devops_console"
version = "2.0.0"
description = "Monitor and manage your devops environment"
authors = ["Croix Bleue <devops@qc.croixbleue.ca>"]
license = "LGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
atlassian-python-api = "^3.26.0"
devops-kubernetes = { path = "_submodules/python-devops-kubernetes", optional = true, develop = true }
devops-sccs = { path = "_submodules/python-devops-sccs", optional = true, develop = true }
fastapi = "^0.79.0"
fastapi-azure-auth = "^3.5.0"
gunicorn = { version = "^20.1.0", optional = true }
hvac = "^0.11.2"
Jinja2 = "^3.1.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = { extras = ["email", "dotenv"], version = "^1.9.1" }
python = "^3.10"
python-multipart = "^0.0.5"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
SQLAlchemy = "^1.4.40"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
mypy = "^0.971"
pytest = { version = "^7.1.2", optional = true }
pytest-cov = { version = "^3.0.0", optional = true }

[tool.poetry.extras]
dev = ["pytest", "pytest-cov"]
local = ["devops-kubernetes", "devops-sccs"]
prod = ["gunicorn", "uvicorn"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
